+++ lending/src/sdk/sdk.ts
@@ -68,8 +68,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "3.3.2";
-    genVersion = "2.131.1";
+    sdkVersion = "3.4.0";
+    genVersion = "2.139.1";
+    userAgent = "speakeasy-sdk/typescript 3.4.0 2.139.1 3.0.0 @codat/lending";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);